Improve view and test to be Object-Oriented
- Move validate_password into SignUpView Class
- Add Validation Class
- Move check_duplicates and generate_jwt into Validation Class
