+++ users/views.py
@@ -1,4 +1,8 @@
 import json
+import jwt
+import re
+
+from datetime               import datetime, timedelta
 
 from django.views           import View
 from django.http            import JsonResponse
@@ -6,11 +10,27 @@ from django.db.utils        import IntegrityError
 from django.forms           import ValidationError
 from django.core.validators import validate_email
 from django.contrib.auth    import login, authenticate
+from django.conf            import settings
 
 from users.models import CustomUser
-from users.utils  import check_duplicate_email, validate_password, generate_jwt
+
+class Validation:
+    def check_duplicate_email(email):
+        if CustomUser.objects.filter(email = email).exists():
+            raise IntegrityError
+
+    def generate_jwt(user):
+        payload      = {'user_id': user.id, 'exp': datetime.now() +timedelta(hours=2)}
+        access_token = jwt.encode(payload, settings.SECRET_KEY, algorithm=settings.ALGORITHM)
+        return access_token
 
 class SignUpView(View):
+    def validate_password(self, password):
+        REGEX_PASSWORD = '^(?=.*\d)(?=.*[a-zA-Z])[0-9a-zA-Z]{8,}$'
+
+        if not re.match(REGEX_PASSWORD, password):
+            raise ValidationError('Enter a valid password.')
+
     def post(self, request):
         try:
             data = json.loads(request.body)
@@ -20,8 +40,8 @@ class SignUpView(View):
             is_doctor       = data['is_doctor']
 
             validate_email(email)
-            validate_password(password)
-            check_duplicate_email(email)
+            self.validate_password(password)
+            Validation.check_duplicate_email(email)
 
             CustomUser.objects.create_user(
                 name      = name,
@@ -40,11 +60,16 @@ class SignUpView(View):
             return JsonResponse({'message' : 'EMAIL_IS_ALREADY_REGISTERED'}, status=400)
 
 class LoginView(View):
+    def generate_jwt(self, user):
+        payload      = {'user_id': user.id, 'exp': datetime.now() +timedelta(hours=2)}
+        access_token = jwt.encode(payload, settings.SECRET_KEY, algorithm=settings.ALGORITHM)
+
+        return access_token
+
     def post(self, request):
         data     = json.loads(request.body)
         email    = data['email']
         password = data['password']
-
         user     = authenticate(request, email=email, password=password)
 
         if user is not None:
@@ -55,7 +80,7 @@ class LoginView(View):
 
                     return JsonResponse({
                         'message'     : 'SUCCESS_PATIENT_LOGIN',
-                        "access_token": generate_jwt(user),
+                        "access_token": Validation.generate_jwt(user),
                         'user_id'     : user.id,
                         'user_name'   : user.name
                     }, status=200)
@@ -69,7 +94,7 @@ class LoginView(View):
 
                 return JsonResponse({
                     'message'     : 'SUCCESS_LOGIN',
-                    "access_token": generate_jwt(user),
+                    "access_token": Validation.generate_jwt(user),
                     'user_id'     : user.id,
                     'user_name'   : user.name
                 }, status=200)
@@ -82,10 +107,10 @@ class LoginView(View):
 class CheckDuplicateEmailView(View):
     def post(self, request):
         try:
-            data  = json.loads(request.body)
+            data = json.loads(request.body)
             email = data['email']
-            check_duplicate_email(email)
-
+            Validation.check_duplicate_email(email)
+            
             return JsonResponse({'message' : 'CAN_REGISTER_WITH_THIS_EMAIL'}, status=201)
         except IntegrityError:
             return JsonResponse({'message' : 'EMAIL_IS_ALREADY_REGISTERED'}, status=400)
